app/requirements.txt
fastapi
fastapi[standard]
uvicorn
SQLAlchemy
psycopg2
python-dotenv
alembic
pydantic[email] 
#hashing            
passlib[bcrypt] 
#jwt equivalent
python-jose
supabase
python-multipart